Add MSA (Min-Surface-Area) random boat placement setting.
Every board config has a maximum possible surface area.  "Surface Area"
being the total number of empty squares adjacent to your boats.  Setting
the "min" surface area to 100 ensures that the random boat plaement
routine will not place any boats adjacent to each other or the endge of
the board.  Setting MSA to 80 will allow a small loss of surface are
(20%) resulting in the ability to have a small amount of boat clustering
or adge placement.  Setting MSA to 0 allows completely random placement
- which is the original logic.