Remove max try counts from random boat placement
Now that configs are rejected if the board isn't large enough to support
100% of max surface area it is safe to remove the arbitrary try counts
from the random boat placement routine.

The random boat placement routine has LOTS of room for improvement, but
if given anough tries it should always be able to find a legal boat
configuration.  Future optimization should attempt to reduce the "try"
set so that duplicate tries are not attempted.  For now it will just
keep trying random placements until it happens upon a pattern that
satisfies the MSA.