Added Hal-9000
Updated some base utility classes as well.  The Throw() class, for
example, has been updated so it no longer throws from its destructor.
And a Msg() class was added to compliment the new Throw() class, as well
as to replace the MSG() inline method.  And the CSV() class has been
split into CSVReader() and CSVWriter().